Apex_Auto_System/requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.1.1
python-dotenv==1.0.0
pymongo==4.6.1
WTForms==3.1.1
email-validator==2.1.0
bcrypt==4.1.2
python-dateutil==2.8.2